requirements.txt
colorama==0.3.3
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.2
psycopg2==2.6
py==1.4.28
py-stackexchange==2.2.4
pytest==2.7.1
selenium==2.46.0
six==1.9.0
static3==0.6.1
waitress==0.8.9
